forgot if for teams migration

+++ database/migrations/2016_95_04_171760_create_lt_Team_table.php
@@ -8,24 +8,27 @@ class CreateLtTeamTable extends Migration
 {
     public function up()
     {
-        Schema::create('team', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('short_id')->unsigned()->nullable();
-            $table->string('name');
-            $table->integer('championship_id')->unsigned()->index(); // A checar
-            $table->string('picture')->nullable();
-            $table->string('entity_type')->nullable(); // Club, Assoc, Fed
-            $table->integer('entity_id')->unsigned()->nullable()->index();
-            $table->timestamps();
+        if (!Schema::hasTable('team')) {
+            Schema::create('team', function (Blueprint $table) {
+                $table->increments('id');
+                $table->integer('short_id')->unsigned()->nullable();
+                $table->string('name');
+                $table->integer('championship_id')->unsigned()->index(); // A checar
+                $table->string('picture')->nullable();
+                $table->string('entity_type')->nullable(); // Club, Assoc, Fed
+                $table->integer('entity_id')->unsigned()->nullable()->index();
+                $table->timestamps();
 
-            $table->foreign('championship_id')
-                ->references('id')
-                ->on('championship')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
+                $table->foreign('championship_id')
+                    ->references('id')
+                    ->on('championship')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
 
-            $table->unique(['championship_id', 'name']);
-        });
+                $table->unique(['championship_id', 'name']);
+            });
+
+        }
     }
 
     public function down()
